refactor: clean dependencies, add more tests, add more targets
