+++ gradle.properties
@@ -10,3 +10,6 @@ android.useAndroidX=true
 # MPP
 kotlin.mpp.androidSourceSetLayoutVersion=2
 kotlin.mpp.enableCInteropCommonization=true
+
+# JS
+org.jetbrains.compose.experimental.jscanvas.enabled=true
