+++ library/core/build.gradle.kts
@@ -8,15 +8,9 @@ kotlin {
     sourceSets {
         commonMain.dependencies {
             implementation(compose.runtime)
-            implementation(compose.material)
             implementation(compose.ui)
-        }
-
-        androidMain.dependencies {
-            // Android
-            implementation(libs.android.core.ktx)
-            implementation(libs.android.lifecycle.runtime)
-            implementation(libs.android.activity.compose)
+            implementation(compose.animation)
+            implementation(compose.foundation)
         }
     }
 }
@@ -29,13 +23,4 @@ android {
     buildFeatures {
         compose = true
     }
-}
-
-dependencies {
-    // Testing
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.android.test.runner)
-    androidTestImplementation(libs.android.espresso)
-    androidTestImplementation(libs.kotlin.coroutines.test)
-    androidTestImplementation(libs.android.test.ext.junit)
 }
