+++ library/ui/src/wasmJsMain/kotlin/com/mr0xf00/easycrop/ui/ImageCropperDialog.wasmJs.kt
@@ -1,0 +1,6 @@
+package com.mr0xf00.easycrop.ui
+
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun isVerticalPickerControls(): Boolean = true
